requirements.txt
apache-airflow==2.7.1
dbt-snowflake==1.5.0
snowflake-connector-python==3.0.0
python-dotenv==1.0.0
pandas>=2.0.0
requests>=2.31.0